data-pjt/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-celery==3.10.0
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.2
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apache-beam==2.48.0
apache-flink==1.20.0
apache-flink-libraries==1.20.0
apispec==6.8.1
argcomplete==3.5.3
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
avro-python3==1.10.2
babel==2.17.0
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
cachelib==0.9.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.1
crcmod==1.7
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==42.0.8
cycler==0.12.1
Deprecated==1.2.18
dill==0.3.1.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
elastic-transport==8.17.1
elasticsearch==8.17.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastavro==1.10.0
fasteners==0.19
feedparser==6.0.11
filelock==3.18.0
find_libpython==0.4.0
findspark==2.0.1
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flower==2.0.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.2.0
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.65.5
gunicorn==23.0.0
h11==0.14.0
hdfs==2.7.3
hf-xet==1.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.31.1
humanize==4.11.0
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
jmespath==0.10.0
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.0.2
kiwisolver==1.4.8
kombu==5.4.2
lazy-object-proxy==1.10.0
limits==4.0.1
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==1.24.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
objsize==0.6.1
openai==1.61.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pemja==0.4.1
pendulum==3.0.0
pillow==11.2.1
pluggy==1.5.0
prison==0.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.1
protobuf==4.25.6
psutil==6.1.1
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyarrow==11.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==1.4.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.12.0
pyparsing==3.2.3
pyspark==3.5.4
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setproctitle==1.3.4
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.8.0
time-machine==2.16.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
